Small number of clusters correction when there is clustering in one treatment group only

This function calculates a small cluster-design adjustment factor which can be used to adjusted effect sizes and their sampling variances from cluster-design studies that adequately handles clustering. This factor can be found in second term of Equation (5) from from Hedges & Citkowicz (2015, p. 1298). The factor in denoted as \(\gamma\) in WWC (2021). The same notion is used here and gave name to the function.

Usage

gamma_1armcluster(N_total, Nc, avg_grp_size, ICC, sqrt = TRUE)

Arguments

N_total

Numerical value indicating the total sample size of the study.

Nc

Numerical value indicating the sample size of the arm/group that does not contain clustering.

avg_grp_size

Numerical value indicating the average cluster size.

ICC

Numerical value indicating the intra-class correlation (ICC) value.

sqrt

Logical indicating if the square root of gamma should be calculated. Default = TRUE

Value

Returns a numerical value for the cluster-design adjustment factor, \(\gamma\).

Details

When calculating effect sizes from cluster-designed studies, it recommended (Hedges, 2007, 2011; Hedges & Citkowitz, 2015; WWC, 2021) to add an adjustment factor, \(\gamma\) to \(d\) whether or not cluster is adequately handled in the studies. Even if clustering is adequately handled, WWC also recommend to use \(\gamma\) as a small number of clusters correction to the variance component. The adjustment factor gamma when there is clustering in one treatment group only is given by

$$\gamma = 1 - \dfrac{(N^C+n-2)\rho}{N_2}$$

where \(N^C\) is the sample size of the group without clustering, \(n\) is the average cluster size, and \(\rho\) is the (often imputed) intraclass correlation.

Then let the naive estimator of Hedges' \(g\) be

$$g_{naive} = J\times \left(\dfrac{\bar{Y}^T_{\bullet\bullet} - \bar{Y}^C_{\bullet}}{S_T} \right)$$

where \(J = 1 - 3/(4df-1)\), \(\bar{Y}^T_{\bullet\bullet}\) it the average treatment effect for the treatment group containing clustering, \(\bar{Y}^C_{\bullet}\) is the average treatment effect for the group without clustering, and \(S_T\) is the standard deviation ignoring clustering. To account for the fact that \(S_T\) systematically underestimates the true standard deviation, \(\sigma_T\), the cluster-adjusted effect size can be obtained from

$$g_T = g_{naive}\sqrt{\gamma}$$

if clustering is properly adjusted for, the sampling variance of \(g_T\) is given by

$$v_{g^T} = \left(\dfrac{1}{N^T} + \dfrac{1}{N^C}\right) \gamma + \dfrac{g^2_T}{2h} $$

where \(N^T\) is the sample size of the treatment group containing clustering and \(h\) is given by

$$ h = \dfrac{[(N-2)(1-\rho) + (N^T-n)\rho]^2} {(N-2)(1-\rho)^2 + (N^T-n)n\rho^2 + 2(N^T-n)(1-\rho)\rho}$$

where \(N\) is the total sample size.

References

Hedges, L. V. (2007). Effect sizes in cluster-randomized designs. Journal of Educational and Behavioral Statistics, 32(4), 341–370. doi:10.3102/1076998606298043

Hedges, L. V. (2011). Effect sizes in three-level cluster-randomized experiments. Journal of Educational and Behavioral Statistics, 36(3), 346–380. doi:10.3102/1076998610376617

Hedges, L. V., & Citkowicz, M (2015). Estimating effect size when there is clustering in one treatment groups. Behavior Research Methods, 47(4), 1295-1308. doi:10.3758/s13428-014-0538-z

What Works Clearinghouse (2021). Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1 Institute of Education Science. https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf

See also

Examples

N <- 100
Nc <- 40
n <- 5
rho <- 0.1

gamma_1armcluster(N_total = N, Nc = Nc, avg_grp_size = n, ICC = rho)
#> [1] 0.978

# Testing function
sqrt(1 - (((Nc + n-2)*rho)/(N-2))) |> round(3)
#> [1] 0.978