Variance calculation when there is clustering in one treatment group only

Insert

Usage

vgt_smd_1armcluster(
  N_cl_grp,
  N_ind_grp,
  avg_grp_size,
  ICC,
  g,
  model = c("Posttest", "reg_coef", "reg_std_coef", "ANCOVA", "DiD"),
  not_cluster_adj = TRUE,
  prepost_cor = NULL,
  F_val = NULL,
  t_val = NULL,
  SE = NULL,
  S = NULL,
  SE_std = NULL,
  R2 = NULL,
  q = NULL
)

Arguments

N_cl_grp

Insert

N_ind_grp

Insert

avg_grp_size

Insert

ICC

Insert

g

Insert

model

Insert

not_cluster_adj

Insert

prepost_cor

Insert

F_val

Insert

t_val

Insert

SE

Insert

S

Insert

SE_std

Insert

Value

A tibble with info ...(Insert)...

Details

Insert

Note

Insert

References

Hedges, L. V., & Citkowicz, M (2015). Estimating effect size when there is clustering in one treatment groups. Behavior Research Methods, 47(4), 1295-1308. doi:10.3758/s13428-014-0538-z

What Works Clearinghouse (2021). Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1 Institute of Education Science. https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf

See also

Examples

vgt_smd_1armcluster(
N_cl_grp = 60, N_ind_grp = 40, avg_grp_size = 10, ICC = 0.1, g = 0.2,
model = "Posttest", not_cluster_adj = TRUE
)
#> # A tibble: 1 × 5
#>      df adj_fct adj_value   V_gt   W_gt
#>   <dbl> <chr>       <dbl>  <dbl>  <dbl>
#> 1  93.0 eta           1.3 0.0544 0.0542