Calculate odds ratios (OR)

This function calculated odds ratios based on various type of input/information, as described in Table 11.10 from Borenstein and Hedges (2019).

Usage

OR_calc(
  A = NULL,
  B = NULL,
  C = NULL,
  D = NULL,
  p1 = NULL,
  p2 = NULL,
  n1 = NULL,
  n2 = NULL,
  OR = NULL,
  LL_OR = NULL,
  UL_OR = NULL,
  SE_OR = NULL,
  V_OR = NULL,
  Z = 1.96,
  add_name_to_vars = NULL,
  vars = dplyr::everything()
)

Arguments

A

Insert

B

Insert

C

Insert

D

Insert

p1

Insert

p2

Insert

n1

Insert

n2

Insert

OR

Insert

LL_OR

Insert

UL_OR

Insert

SE_OR

Insert

V_OR

Insert

Z

Insert

add_name_to_vars

Optional character string to be added to the variables names of the generated tibble.

vars

Variables to be reported. Default is NULL. See Value section for further details.

Value

A tibble with information about OR, OR_LN, vln_OR.

Details

Add details

Event	Non-event	Total

Treated		A	B	n1

Control		C	D	n2

Note

Read Borenstein and Hedges (2019) for further details.

References

Borenstein and Hedges (2019). Effect sizes for meta-analysis. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (3rd ed., pp. 207–242). Russell Sage Foundation West Sussex.

Examples

# Using raw events
OR_calc(A = 20, B = 80, C = 10, D = 90)
#> # A tibble: 1 × 3
#>      OR ln_OR vln_OR
#>   <dbl> <dbl>  <dbl>
#> 1  2.25 0.811  0.174

# Using proportions
OR_calc(p1 = .53, p2 = .11, n1 = 20, n2 = 26)
#> # A tibble: 1 × 3
#>      OR ln_OR vln_OR
#>   <dbl> <dbl>  <dbl>
#> 1  9.12  2.21  0.594

# Using raw OR and CI
OR_calc(OR = 2.25, LL_OR = 1.5, UL_OR = 3)
#> # A tibble: 1 × 3
#>      OR ln_OR vln_OR
#>   <dbl> <dbl>  <dbl>
#> 1  2.25 0.811 0.0313

# Adding suffix to variables and selecting specific variables
OR_calc(A = 20, B = 80, C = 10, D = 90, add_name_to_vars = "_test", vars = OR_test)
#> # A tibble: 1 × 1
#>   OR_test
#>     <dbl>
#> 1    2.25