Variance calculation when there is clustering in one treatment group only

Insert

Usage

vgt_smd_1armcluster(
  N_cl_grp,
  N_ind_grp,
  avg_grp_size,
  ICC,
  g,
  model = c("Posttest", "ANCOVA", "DiD", "Reg_coef", "Std_reg_coef"),
  not_cluster_adj = TRUE,
  prepost_cor = NULL,
  F_val = NULL,
  t_val = NULL,
  SE = NULL,
  SD = NULL,
  SE_std = NULL,
  R2 = NULL,
  q = 1
)

Arguments

N_cl_grp

Insert

N_ind_grp

Insert

avg_grp_size

Insert

ICC

Insert

g

Insert

model

Insert

not_cluster_adj

Insert

prepost_cor

Insert

F_val

Insert

t_val

Insert

SE

Insert

SD

Insert

SE_std

Insert

R2

Insert

q

Insert

Value

A tibble with info ...(Insert)...

Details

Table 1 illustrates all cluster adjustment of variance estimates from pre-test and/or covariate adjusted measures that can be calculated with the vgt_smd_1armcluster()

Table 1
Sampling variance estimates for \(g_T\) across various models for handling cluster, estimation techniques, and reported quantities.

Calculation type/
reported quantities	Cluster-adjusted (model)
sampling variance	Not cluster-adjusted (model)
sampling variance

ANCOVA, adj. means
\(R^2, N^T, N^C\)	\((1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2(h-q)}.\)	\((1-R^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2(h-q)}.\)

ANCOVA, adj. means
\(R^2_{imputed}, N^T, N^C\)	\((1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2(h-q)}.\)	\((1-0^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2(h-q)}.\)

ANCOVA, adj. means
\(F, (t^2), N^T, N^C\)	\(\left(\frac{g^2_T}{F}\right) \gamma + \frac{g^2_T}{2(h-q)}.\)	\(\left(\frac{g^2_T}{F}\right) \eta + \frac{g^2_T}{2(h-q)}.\)

ANCOVA, pretest only
\(F, (t^2), N^T, N^C\)	\(\left(\frac{g^2_T}{F}\right) \gamma + \frac{g^2_T}{2h}.\)	\(\left(\frac{g^2_T}{F}\right) \eta + \frac{g^2_T}{2h}.\)

ANCOVA, pretest only
\(r, N^T, N^C\)	\((1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.\)	\((1-r^2) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.\)

Reg coef
\(SE, S_T, N^T, N^C\)	\(\left(\frac{SE}{S_T}\right)^2 \gamma + \frac{g^2_T}{2(h-q)}.\)	\(\left(\frac{SE}{S_T}\right)^2 \eta + \frac{g^2_T}{2(h-q)}.\)

Reg coef, pretest only
\(SE, S_T, N^T, N^C\)	\(\left(\frac{SE}{S_T}\right)^2 \gamma + \frac{g^2_T}{2h}.\)	\(\left(\frac{SE}{S_T}\right)^2 \eta + \frac{g^2_T}{2h}.\)

Std. reg coef
\(SE_{std}, N^T, N^C\)	\(SE^2_{std} \gamma + \frac{g^2_T}{2(h-q)}.\)	\(SE^2_{std} \eta + \frac{g^2_T}{2(h-q)}.\)

Std. reg coef, pretest only
\(SE_{std}, N^T, N^C\)	\(SE^2_{std} \gamma + \frac{g^2_T}{2h}.\)	\(SE^2_{std} \eta + \frac{g^2_T}{2h}.\)

DiD, gain scores
\(r, N^T, N^C\)	\(2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.\)	\(2(1-r) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.\)

DiD, gain scores
\(r_{imputed}, N^T, N^C\)	\(2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \gamma + \frac{g^2_T}{2h}.\)	\(2(1-.5) \left(\frac{1}{N^T} + \frac{1}{N^C}\right) \eta + \frac{g^2_T}{2h}.\)

DiD, gain scores
\(t, N^T, N^C\)	\(\left(\frac{g^2}{t^2}\right) \gamma + \frac{g^2_T}{2h}.\)	\(\left(\frac{g^2}{t^2}\right) \eta + \frac{g^2_T}{2h}.\)

Note: \(R^2\) "is the multiple correlation between the covariates and the outcome" (WWC, 2021), \(\eta = 1 - (N^C+n-2)\rho/(N-2)\), see eta_1armcluster, \(\gamma = 1 - (N^C+n-2)\rho/(N-2)\), see eta_1armcluster, \(r\) is the pre-posttest correlation, and \(q\) is the number of covariates. Std. = standardized.

"It is often desired in practice to adjust for multiple baseline characteristics. The problem of \(q\) covariates is a straightforward extension of the single covariate case (...): The correlation coefficient estimate \(r\) is now obtained by taking the square root of the coefficient of multiple determination, \(R^2\)" (Hedges et al. 2023, p. 17) and \(df = h-q\).

Note

Insert

References

Hedges, L. V., & Citkowicz, M (2015). Estimating effect size when there is clustering in one treatment groups. Behavior Research Methods, 47(4), 1295-1308. doi:10.3758/s13428-014-0538-z

What Works Clearinghouse (2021). Supplement document for Appendix E and the What Works Clearinghouse procedures handbook, version 4.1 Institute of Education Science. https://ies.ed.gov/ncee/wwc/Docs/referenceresources/WWC-41-Supplement-508_09212020.pdf

See also

Examples

vgt_smd_1armcluster(
N_cl_grp = 60, N_ind_grp = 40, avg_grp_size = 10, ICC = 0.1, g = 0.2,
model = "ANCOVA", not_cluster_adj = TRUE, R2 = 0.5, q = 3
)
#> # A tibble: 1 × 8
#>       h    df var_term1 adj_fct adj_value   V_gt   W_gt Wgt_test
#>   <dbl> <dbl>     <dbl> <chr>       <dbl>  <dbl>  <dbl>    <dbl>
#> 1  93.0  90.0    0.0208 eta           1.3 0.0273 0.0271   0.0271